Add storage file, decouple persistence from page_generator
`page_generator` is no longer a class and now it's just a single function

Added storage, a wrapper aroung a python dictionary, with some simple
add / remove / contains functions. Automatic id with uuid.
