+++ server/server.py
@@ -3,6 +3,7 @@
 from bottle import run, abort, post, get, request, HTTPResponse
 
 import config
+import storage
 import page_generator
 
 
@@ -13,9 +14,10 @@ def process_help():
     media_type = 'image'
     media_url = result['media']['content']
 
-    page_id = page_generator.generate_page(media_url, media_type)
-    user_url = config.get_user_endpoint() + page_id
+    page_content = page_generator.generate_page(media_url, media_type)
+    page_id = storage.save_page(page_content)
 
+    user_url = config.get_user_endpoint() + page_id
     done_url = config.get_done_url()
 
     return {
@@ -26,10 +28,10 @@ def process_help():
 
 @get('/resolve/<page_id>')
 def serve_static_page(page_id):
-    if not page_generator.has_generated(page_id):
+    if not storage.contains(page_id):
         abort(404, 'Not found')
 
-    return page_generator.retrieve_page(page_id)
+    return storage.get_page(page_id)
 
 
 @post('/api/done')
@@ -51,14 +53,13 @@ def stop_help():
 
     page_id = user_url.split('/')[-1]
 
-    if not page_generator.has_generated(page_id):
+    if not storage.contains(page_id):
         return HTTPResponse(status=403, body="Auth URL doesn't match!")
 
-    page_generator.remove_page(page_id)
+    storage.remove_page(page_id)
     return HTTPResponse(status=200)
 
 
 if __name__ == "__main__":
-    page_generator = page_generator.PageGenerator()
-
+    storage = storage.Storage()
     run(host=config.get_domain_name(), port=config.get_domain_port(), debug=True)
